Pull server.properties config into its own class
