+++ src/main/java/co/aikar/timings/TimingsExport.java
@@ -78,7 +78,7 @@ public class TimingsExport extends Thread {
         if (!Timings.isPrivacy()) {
             out.put("server", Server.getInstance().getName());
             out.put("motd", Server.getInstance().getMotd());
-            out.put("online-mode", Server.getInstance().getServerConfig().getXboxAuth());
+            out.put("online-mode", Server.getInstance().getServerConfig().getPropertiesConfig().getXboxAuth());
             out.put("icon", ""); //"data:image/png;base64,"
         }
