+++ src/main/java/org/cloudburstmc/server/command/defaults/DefaultGamemodeCommand.java
@@ -39,7 +39,7 @@ public class DefaultGamemodeCommand extends Command {
         try {
             GameMode gameMode = GameMode.from(args[0].toLowerCase());
 
-            sender.getServer().getServerConfig().setGamemode(gameMode);
+            sender.getServer().getServerConfig().getPropertiesConfig().setGamemode(gameMode);
             sender.sendMessage(new TranslationContainer("commands.defaultgamemode.success", gameMode.getTranslation()));
         } catch (IllegalArgumentException e) {
             sender.sendMessage("Unknown game mode"); //TODO: translate?
